Change argument nBytes in write() and read() functions from byte data type to unsigned int. This allows devices with page size of 256 bytes or larger. Fixes #1.

+++ extEEPROM.h
@@ -82,9 +82,9 @@ class extEEPROM
     public:
         extEEPROM(eeprom_size_t deviceCapacity, byte nDevice, unsigned int pageSize, byte eepromAddr = 0x50);
         byte begin(twiClockFreq_t twiFreq = twiClock100kHz);
-        byte write(unsigned long addr, byte *values, byte nBytes);
+        byte write(unsigned long addr, byte *values, unsigned int nBytes);
         byte write(unsigned long addr, byte value);
-        byte read(unsigned long addr, byte *values, byte nBytes);
+        byte read(unsigned long addr, byte *values, unsigned int nBytes);
         int read(unsigned long addr);
 
     private:
